usb: gadget: u_ether: Make sure mac address gets set if is is passed to gether_setup_name
The android gadget passes the USB mac to gether_setup_name. Now it is set correctly in that
function so no random mac addresses are set. This should make connecting through
usb a lot nicer, as the same network config can be used. The mac will now not change
between reboots and replugs.

[sbj] Fix usb host addr so we have a fixed mac for USB network connections.

Signed-off-by: Philippe De Swert <philippe.deswert@jollamobile.com>