drm/i915: Split set_display into val/calc and commit
BZ: 1265

As a preparation to "true" display atomic update, split the display
change analysis and actual write into hardware into different logical
sections
RRB2 is taken care as part of the flip operation.
Z-order is calculated and the cached in the register, written at the
begining of the commit and in the later places where sprite control
register is touched the first 3 bits related to z-order are non-masked.

merge from bbb254d761b772bc8cfe02910e1dccdb719e0aa0 for Gerrit 280612

Iorig-Change-Id: I554da597e40db2a4790a7925030d16c66746e669
Change-Id: Iad0ff00f1d78769fed01883262ac1745b45eb993
Reviewed-on: http://192.168.1.79/gerrit/9621
Reviewed-by: b532 <b532@borqs.com>