Make sure we always have a valid USB MAC
Seems that by default the mac is set to 00:00:00:00:00:00. However
this cannot be used and thus unless we set the mac in userspace, this
wrong one will be used. Now we make sure that the set mac is checked for
validity and then make sure it is update in the gadget driver.

Signed-off-by: Philippe De Swert <philippe.deswert@jollamobile.com>