gpu: i915: Don't toggle tc358860 state with runtime pm.
The tc358860 converter should be enabled when display is on and
disabled when display is off. There is no need to toggle state also
with runtime pm, which is causing an extra disable-enable cycle when
device is resuming from suspend. Removed tc358860_enable/disable
calls from i915 runtime suspend/resume calls.

[display] Don't toggle tc358860 state with runtime pm. Contributes to JB#31051

Signed-off-by: Kalle Jokiniemi <kalle.jokiniemi@jolla.com>