gpu: i915: Reduce tc358860_read_reg latency.
tc358860_read_reg and tc358860_read_reg1 showed up in resume time
optimization as one of the main source of latency. The main reason
seems to be the used msleep(2) calls that turn into 20ms delays
on Baytrail platform.

Changed the msleeps into usleep_range. The range was set from 1ms to 2ms,
and seems to provide same functionality as before with much less latency.

[display] Reduce tc358860_read_reg latency. Contributes to JB#31051

Signed-off-by: Kalle Jokiniemi <kalle.jokiniemi@jolla.com>