[usb] Go one power class up. Contributes: JB#33252
Since the tablet seems not to actually draw the announced 500mA as
it should, switching to the next power level supported by the
charging chip. This should be anyways ok for USB3 ports. And when
measured does not seem to overdraw current dramatically. It does
improve the charging behaviour a lot.

Signed-off-by: Philippe De Swert <philippe.deswert@jollamobile.com>